+++ src/img/zrandom.cpp
@@ -12,9 +12,9 @@
 //  For everything else
 static unsigned long long rdtsc()
 {
-  unsigned int lo, hi;
+  uint32_t lo, hi;
   __asm__ __volatile__ ("rdtsc" : "=a" (lo), "=d" (hi));
-  return (static_cast<unsigned long long>(hi) << 32) | lo;
+  return (static_cast<uint64_t>(hi) << 32) | lo;
 }
 
 #endif
